[librpsecure] os-secure.h: Reduce the syscall table from `int[]` to `int16_t[]`.
There's nowhere near even 1,000 syscalls on each Linux architecture,
so we can cut the size of the libseccomp whitelist tables in half.

NOTE: 32-bit ARM does have some special "ARM_*" syscalls starting at
983041. This might only be for Chrome OS. Either way, these syscalls
aren't relevant for rom-properties, so they can stay disabled.

+++ src/amiibo-data/amiiboc.cpp
@@ -129,7 +129,7 @@ static int set_security_options(void)
 #if defined(_WIN32)
 	param.bHighSec = 1; // Disable Win32k syscalls. (NTUser/GDI)
 #elif defined(HAVE_SECCOMP)
-	static constexpr int syscall_wl[] = {
+	static constexpr int16_t syscall_wl[] = {
 		// Syscalls used by amiiboc.
 		SCMP_SYS(close),
 		SCMP_SYS(gettimeofday), // 32-bit only?
